Fixed error if SearchTeam on a team with no data
In SearchTeam, added a check to see if the team being searched for has
data, if not, prints out a "no data" line and returns.  For the file
write situation, if multiple consecutive no data teams are present,
script errors out if you don't set the index before returning.  Not sure
why, but set it anyway.  Also removed return in MatchReport so that data
for all upcoming matches are written to the file, not just the first
one, now that this won't error out anymore on our test dataset.